Return Tobasco._recordSubmission() to original design
Saving blockhashes was added to make it easier to query blockhashes > 255 blocks old when slashing and was used in the Burn._verifyGatewaySignatures(). However, if you are slashing because of a missed ToB then _recordSubmission() will not have been called and the blockhash not saved making the change useless. It's been reverted back to just saving (blocknumber -> bool)