Refactor (#2)
* separate slashing/ownership logic from tobasco

* fix tests

* Add slasher contract that follows the fault attribution model in the constraints API

https://github.com/eth-fabric/constraints-specs/tree/fault-attribution/specs

* add logic for Gateway to vindicate themselves via MIP

* comment out for now

* fix off by one error

* test infra in place

* Return Tobasco._recordSubmission() to original design

Saving blockhashes was added to make it easier to query blockhashes > 255 blocks old when slashing and was used in the Burn._verifyGatewaySignatures(). However, if you are slashing because of a missed ToB then _recordSubmission() will not have been called and the blockhash not saved making the change useless. It's been reverted back to just saving (blocknumber -> bool)

* add tests for markGatewaySlashable() and slash()

* use blockNumber instead of blockTimestamp

* Update README.md

* Update Tobasco.s.sol